+++ docs/tutorials/pisa/settings.py
@@ -2,5 +2,4 @@ from lino.projects.std.settings import *
 SITE = Site(
     globals(),
     ['lino.modlib.system', 'lino.modlib.users', 'pisa'],
-    languages='en de fr',
-    user_model='users.User')
+    languages='en de fr')