+++ lino/modlib/users/__init__.py
@@ -1,8 +1,8 @@
-# Copyright 2011-2014 Luc Saffre
+# Copyright 2011-2015 Luc Saffre
 # License: BSD (see file COPYING for details)
 
-"""Lino's :mod:`lino.modlib.users` is an alternative to Django's
-:mod:`django.contrib.auth` module.
+"""Lino's :mod:`lino.modlib.users` is Lino's replacement for Django's
+:mod:`django.contrib.auth` module. See also :doc:`/dev/users`.
 
 .. autosummary::
    :toctree:
@@ -15,21 +15,8 @@
     fixtures.demo2
 
 
-This module is much more simple and does not require
-:mod:`django.contrib.sessions` to be installed.
-See :srcref:`docs/tickets/31` for discussion.
-
-To use it, you must override :setting:`user_model` and
-:setting:`get_installed_apps` in your
-:class:`lino.lino_site.Site`. Example::
-
-    user_model = 'users.User'
-    
-    def get_installed_apps(self):
-        yield super(Lino,self).get_installed_apps()
-        yield 'lino.modlib.system'
-        yield 'lino.modlib.users'
-        # continue with your own modules
+This module does not require :mod:`django.contrib .sessions` to be
+installed. See :srcref:`docs/tickets/31` for discussion.
 
 """
 
@@ -42,3 +29,6 @@ class Plugin(ad.Plugin):
     "Ceci n'est pas une documentation."
     verbose_name = _("Users")
 
+    def on_init(self):
+        super(Plugin, self).on_init()
+        self.site.set_user_model('users.User')