+++ lino/projects/presto/settings/doctests.py
@@ -1,3 +1,3 @@
-from ..settings import *
+from lino.projects.presto.settings import *
 SITE = Site(globals(), no_local=True, remote_user_header='REMOTE_USER')
 DEBUG = True